game is working, TODO: instructions and count need to be added
icons are stored in an array, based on the number of grids, a slice of the array is created for the game. Colors are taken randomly from an array and added to icons. Each item exists in an object within an array with a key, the key is used as a data id to identify to identical icons when playing the game